todoHero/pyproject.toml
[tool.poetry]
name = "todohero"
version = "0.1.0"
description = ""
authors = ["Samuel Owen <SamHirael@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
ruff = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
